+++ commands/spigot-platform/src/test/java/net/apartium/cocoabeans/commands/spigot/CommandsSpigotTestBase.java
@@ -42,6 +42,15 @@ public class CommandsSpigotTestBase {
         execute(sender, split[0], Arrays.copyOfRange(split, 1, split.length));
     }
 
+    public void execute(CommandSender sender, String commandName, String args) {
+        try {
+            commandManager.handle(new SpigotSender<>(sender), commandName, args.split("\\s+"));
+        } catch (Throwable e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+
     public void execute(CommandSender sender, String commandName, String[] args) {
         try {
             commandManager.handle(new SpigotSender<>(sender), commandName, args);
@@ -50,7 +59,7 @@ public class CommandsSpigotTestBase {
         }
     }
 
-    List<String> evaluateTabCompletion(CommandSender sender, String label, String args) {
+    public List<String> evaluateTabCompletion(CommandSender sender, String label, String args) {
         return commandManager.handleTabComplete(new SpigotSender<>(sender), label, args.split("\\s+"));
     }